requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Machine Learning and Computer Vision
torch==2.6.0
torchvision==0.21.0
clip-by-openai==1.1
opencv-python==4.8.1.78
scikit-image==0.21.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3

# Image Processing
Pillow==10.1.0
imagehash==4.3.1

# Geolocation and Mapping
geopy==2.4.1
astral==3.2
timezonefinder==6.2.0
pytz==2023.3
python-dateutil==2.8.2
shapely==2.0.2

# HTTP and API
requests==2.31.0
httpx==0.25.2

# Authentication and Security
PyJWT==2.8.0
cryptography==41.0.7
firebase-admin==6.4.0

# Database
# sqlite3 is built-in with Python

# Configuration and Environment
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.7.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: Compression
gzip  # Built-in with Python